+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- **breaking** [node] fix: Events created from exception shouldnt have top-level message attribute
+- [utils] ref: Update wrap method to hide internal sentry flags
+- [utils] fix: Make internal Sentry flags non-enumerable in fill util
+
 ## 4.5.3
 
 - [browser]: fix: Fix UnhandledPromise: [object Object]